LayoutTests/system-preview/rel-ar-setting-disabled-expected.txt


PASS a.relList.supports("ar") should be false when SystemPreviewEnabled is false.
